+++ libs/core/src/index.ts
@@ -6,9 +6,12 @@ export * from './promise';
 export * from './search';
 export * from './sleep';
 export * from './models/Base';
-export { AUTH_USER_CACHE_KEY } from './constants';
+export { AUTH_ADMIN_CACHE_KEY } from './constants';
 export { exportToExcel, exportToCsv } from './export.utils';
 export {
   getBreadcrumbItemsFromMap,
   type BreadcrumbItemFromMap,
 } from './breadcrumb.utils';
+
+export * from './configs/rbac.config';
+export * from './configs/auth.config';