+++ libs/features/admin/auth/src/pages/Login.tsx
@@ -1,9 +1,9 @@
-import { cache, AUTH_USER_CACHE_KEY } from '@core';
+import { cache, AUTH_ADMIN_CACHE_KEY } from '@core';
 import { useNavigate } from 'react-router';
 import { useLoading } from '../contexts/LoadingContext';
 import { LoginForm, type LoginFormValues } from '@ui';
 import { getRolesFromBackend } from '../services/auth.service';
-import { getPermissionsForRoles } from '../config/rbac.config';
+import { getPermissionsForRoles } from '@core';
 import useAuthStore from '../store/useAuthStore';
 
 export default function Login() {
@@ -16,7 +16,7 @@ export default function Login() {
     try {
       const roles = await getRolesFromBackend(values.username);
       const permissions = getPermissionsForRoles(roles);
-      cache.setCache(AUTH_USER_CACHE_KEY, {
+      cache.setCache(AUTH_ADMIN_CACHE_KEY, {
         token: 'demo-token',
         username: values.username,
         remember: values.remember,